Update support.md
Critical just makes sense to an English first reader. Not including it means that "everything and anything may or may not be critical". 

Therefore I'm leaving it in there. If you'd like to decline the PR, fine. 

Arguing with you over the sentence structure of an English targeted product isn't worth my time.