feat(xo-server/backups-ng): limit number of gc-ed deltas
Merging multiple VHDs is currently a slow process which could be optimized by doing a single merge of a synthetic delta.

Until this is implement, the number of garbage collected deltas should be limited to avoid taking too much time (and possibly interupted jobs) in case of an important retention change.