fix(proxy): forbid local remotes
These remotes are local to the proxy appliance and not the XOA, this might be surprising.

Also, proxy appliances are supposed to be stateless which is not the case with local remotes.

We'll revert this in the future if a valid use case is presented.

These remotes are not disabled in dev mode because they are used for tests, but that might change in the future.