feat(xo-server,xo-web): split XO Proxy upgrade and force upgrade workflow
The normal upgrade is using the API call, the force upgrade is using the reboot workflow.

Also, we should not fallback to the reboot workflow automatically on API call failure
because it can be normal due to xoa-updater and/or XO Proxy being rebooted and rebooting
the VM may break the current upgrade and leave npm in an broken state.