docs(users): SAML instructions for use with Google Workspace (#7846)
This shows a working example of how to connect XO SAML-plugin with Google Workspace SAML.

It's a rough example, with simple screenshots. And I've mentioned the caveat that a successfull authentication points the user to the login screen. Needing to browse into the bare URL, for access into the dahsboard. But at least it works!