airgap.md: Explain how to prepare XCP-ng for air-gapped systems
This page previously explained how to prepare Xen Orchestra for air-gapped environments.
Now, we want to explain the same thing for both XCP-ng and Xen Orchestra, in the same page.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>

airgap.md: Explain how to prepare XCP-ng for air-gapped systems

This page previously explained how to prepare Xen Orchestra for air-gapped environments.
Now, we want to explain the same thing for both XCP-ng and Xen Orchestra, in the same page.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>