airgap.md: Optimize steps to prepare XOA for air-gap systems
This article now covers both XCP-ng and Xen Orchestra.

As a result, the text is now streamlined across the whole page,
for an easier reading exprience and a more coherent style between the two parts.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>

airgap.md: Optimize steps to prepare XOA for air-gap systems

This article now covers both XCP-ng and Xen Orchestra.

As a result, the text is now streamlined across the whole page,
for an easier reading exprience and a more coherent style between the two parts.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>