Update users.md with more usefull text for Google SAML
I noticed that the SAML-docs I wrote, could use some tweaking. To make it more helpfull. So I added a few points to better point to where to find this inside Googles admin dashboard.

Also. Added a small note about a potential certificate error, if there isn't a newline at the bottom of the certificate field.