doc(XO): Add link to the V2V section of the XCP-ng doc
The XCP-ng doc has been recently updated with more details
on how to migrate from VMware to a Vates environment,
with a choice of methods including V2V within Xen Orchestra.
As a result, a mention of V2V is added to the XO doc, with a link
to the relevant section in the XCP-ng doc.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>