doc(XO): Update URL of the Incremental replication page
Continuous replication was renamed as incremental replication.
The relevant page was already renamed but not the URL.
Also, a duplicate page with the correct URL already existed but was not linked in the navigation bar.
The Continuous replication page is now deleted and
the Incremental replication page is now linked in the nav bar.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>