doc(XO): Correct external links to the XO doc and pricing page
On various pages of the documentation, links to other pages followed the old syntax. External links have been replaced with internal links where appropriate, and the other links were updated with the current syntax.

In addition, links to the pricing page now point to the current version instead of the old pricing page

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>