doc(XO): Fix broken links to plugin documentation
Previously, the plugin documentation resided on a standalone page that wasn't listed in the sidebar and duplicated content from the Architecture page's plugin section.
This standalone Plugins page was removed in a previous commit (https://github.com/vatesfr/xen-orchestra/pull/8539), causing several links to break.
This commit updates those links to point to the correct location.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>