docs(manage_infrastructure): document software RAID health check (#8576)
While Software RAID is provided as-is, Xen Orchestra allows users to monitor the health of their software RAID arrays.

This includes a dedicated status interface and alert messages for any issues.

This information was previously documented in the Xen Orchestra 5.102 release blog post, and is now integrated into the XO technical documentation.

This update ensures users have a single reference point for technical information, eliminating the need to search through blog posts.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>