docs(advanced, manage_infrastructure): deduplicate alerts content
The Alerts page was not listed in the sidebar and duplicated content from the Alerts section in the Advanced page. This commit removes the standalone Alerts page and redirects all links pointing to it to the Alerts section of the Advanced page. This change also fixes a broken anchor warning in Docusaurus, which could prevent new documentation builds from succeeding.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>