docs(usage-reports): fix link to the usage report plugin documentation
Previously, the Readme and Usage files related to the server-usage-report plugin contained links pointing to a general page on plugins.

The server-usage-report plugin has since been documented in the Management → Manage infrastructure management page of the Xen Orchestra documentation. As a result, these links have become obsolete.

This commit make the obsolete links point to the actual, new documentation on the server-usage-report plugin.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>