docs(plugins): fix links to various plugins
The documentation for mutliple plugins was updated or moved outside the general Plugins page over time.

As a result, the links contained in various plugin's Readme or Usage files got obsolete, as they all point to the general Plugins page, instead of the specific plugin documentations.

This commit updates those links with the relevant location.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>