docs(restapi): clean up the rest api page
As the API endpoints are getting more and more documented in Swagger, there are now multiple sources of documentation of the REST API, which can cause confusion for users.

This improvement removes references to API endpoints from the XO technical documentation so that users are encouraged to read about them directly on Swagger, which should act as the single point of reference.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>