feat(xo-web/user): add XO6 only languages to dropdown
XO6 is translated into languages, not existing for XO5. And as it is not possible to select language in XO6 yet (is inherited from XO5 settings), interim solution could be to add these new languages into XO5 user settings, to make them actually accessible in XO6 (without need to edit this mimified index.js file in XO installation or masking as file of existing language file before compilation from sources). It will not affect XO5, because it will simply fallback into English in case of not existing language.