docs(xo): remove obsolete administration page
The Administration page duplicates content in the VMs section of the Infrastructure management page.

Although the Administration page seemingly refers to itself as a 'section', it contains no subsections or extra content, so this might as been a draft for a bigger page, and was never elaborated upon.

In any case, this page isn't accessible from the navigation bar and doesn't seem to add much value to the doc, so I suggest we remove it from the tree view.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>