docs(xo): deduplicate backups and disaster recovery content
The 'Backups and Disaster Recovery' page in the XO documentation wasn't listed in the navigation bar and had content that was already present in other pages in the doc.

The last section of that page dealt with scheduled snapshots and completes the existing content on 'rolling snapshots'. That content has been added to the rolling snapshots section, so that everything can be grouped in a single location.

Finally, since the Backups and disaster recovery page is now obsolete, it has been deleted.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>