doc(xo): deduplicate content across multiple pages (#8851)
* docs(xo): deduplicate acls content

The ACLs page in the XO documentation wasn't listed in the navigation bar and had content that duplicated the ACLs section of the Users page.

This commit removes the standalone page and updates the valid section with minor improvements (presentation, some missing info from the lone ACLs page) so that it becomes the only source of information on the topic.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>

* docs(xo): remove obsolete administration page

The Administration page duplicates content in the VMs section of the Infrastructure management page.

Although the Administration page seemingly refers to itself as a 'section', it contains no subsections or extra content, so this might as been a draft for a bigger page, and was never elaborated upon.

In any case, this page isn't accessible from the navigation bar and doesn't seem to add much value to the doc, so I suggest we remove it from the tree view.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>

* docs(xo): deduplicate backups and disaster recovery content

The 'Backups and Disaster Recovery' page in the XO documentation wasn't listed in the navigation bar and had content that was already present in other pages in the doc.

The last section of that page dealt with scheduled snapshots and completes the existing content on 'rolling snapshots'. That content has been added to the rolling snapshots section, so that everything can be grouped in a single location.

Finally, since the Backups and disaster recovery page is now obsolete, it has been deleted.

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>

* Update docs/docs/users.md

Co-authored-by: Laure S. <laure.souai@vates.tech>

---------

Signed-off-by: Thomas Moraine <thomas.moraine@vates.tech>
Co-authored-by: Laure S. <laure.souai@vates.tech>